ahp有问题，注释掉的都未删，目前只用2alg测试

+++ src/main/resources/sql/algresult.sql
@@ -5,8 +5,8 @@ CREATE TABLE `algresult` (
   `releaseId` int(11) DEFAULT NULL COMMENT '发布任务ID，假设本次发布有100个input',
   `algName1` varchar(45) DEFAULT NULL COMMENT '算法1名称',
   `algName2` varchar(45) DEFAULT NULL,
-  `score1` float DEFAULT NULL COMMENT 'Alg1的得分',
-  `score2` float DEFAULT NULL,
+  `score1` double DEFAULT NULL COMMENT 'Alg1的得分',
+  `score2` double DEFAULT NULL,
   `winAlgname` varchar(45) DEFAULT NULL COMMENT '所有input的AB最后赢的算法名称',
   `iffinal` varchar(45) DEFAULT NULL,
   PRIMARY KEY (`id`)
